input: misc: qpnp-power-on: Add static inline stubs for optional APIs
In change 5954080 ("soc: qcom: Add support of reboot reason driver"),
the qcom-reboot-reason driver includes qpnp-power-on.h and uses
qpnp_pon_set_restart_reason() as a fallback when NVMEM support is
absent.

On targets where CONFIG_INPUT_QPNP_POWER_ON is not enabled, this
causes "unused function" build warnings due to unreferenced
declarations.

To handle this cleanly, static inline are added for relevant APIs
under the #else section in qpnp-power-on.h. This ensures safe
inclusion of the header regardless of config state.

Change-Id: Icb3e66f7166135bbe2c931194786678c31a75af5
Signed-off-by: Mohammed Aakib Patel <quic_maiyubpa@quicinc.com>